refactor(scout-for-lol): extract useExploreConversation to stay under the line limit
Why:
Rebasing onto main picked up further churn, and the explore.tsx transcript-
failure fix from the previous commit pushed the file to 505 lines against the
500-line max.

What:
`useExploreConversation` — the data hook that projects both explore queries
through `Loaded` and is not the route component itself — moves to
`#src/hooks/use-explore-conversation.ts`, matching the existing
`use-explore-turn-actions.ts` convention. It was already fully self-contained
(only `useTRPC`, `useQuery`, and `Loaded`), so the move is mechanical.

Verification:
  bunx turbo run typecheck lint test build --filter=@scout-for-lol/app
    — passing; check-architecture clean over 314 modules
